apps/worker/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
Pillow>=10.0.0
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
trl>=0.8.0
accelerate>=0.28.0
datasets>=2.18.0
